+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 1.1.15
+- Fix #62: internal IdleStop message for the ProcessInstance actor is now configured to be serialized by Kryo
 - Fix #59: disabled the usage of Akka Distributed Data until the growing memory issue in the shared process metadata feature is solved.
 - Fix #57: configured the default actor idle-timeout as 5 minutes (baker.actor.idle-timeout config can be overridden in the application.conf)
 - Fix #56: fixed one unhandled message warning for ProcessInstanceEvent